Testing gen pub -e -p -q.
Fixing public key generation, which segfaulted with the couple <p, q>, and
adding a small test to asssert it.
Moreover, changing #!/bin/bash to #!/bin/sh. Apparently using bash is almost
always wrong: may be outdated in new osx, not present in some linux, and
generally not really needed in comparaison with bash.