Using a smaller dataset of primes, if not in cluster mode.
Refine the script primes.sh to accept a parameter $1, the number of millions of
primes that wants to be used as pool.
Now, by default a simple user downloads only the first million of primes.
A cluster user instead will download 20M primes.
Please note that the maximum is 22 as of today.